backend/scrape-server/requirements.txt
flask
beautifulsoup4
pytest-playwright
google-cloud-aiplatform
python-dotenv

